only do one copy for BLOB values
(was two before; now we just copy from SQLite to rusqlite and then
reference that and make an Erlang binary pointing to the rusqlite
value).

In the process `XqliteVal` was deemed no longer necessary because we had
to fragment the encoding logic. Thus, having an `impl Encoder` for one
object but still pattern-match on rusqlite Value before invoking the
encoding implementation became a hurdle. This is addressed by making a
simple non-trait-bound function that encodes all possible SQLite values
to Erlang terms.