Full rework (#9)
* First rough r2d2 integration in the Rust code

* use a new attribute macro for `resource`

Co-authored-by: Benedikt Reinartz <filmor@gmail.com>

* also import the resource attribute macro

* simplify return values handling

* do not manually register a resource if using `resource_impl`

* simplify return values by removing XqliteResult

* simplify return values by removing XqliteOk

* use explicit `.encode(env)` for better return value control

* remove unnecessary `impl RefUnwindSafe`

* avoid lifetime issues via stringifying errors and

a big refactor to accomodate for it. Also utilize `From`, `?` and normal
error handling.

* remove nif_version_* requirement

* add more error variants for clarity

* remove unneeded Emacs gitignore rule

* add support for writing PRAGMAs

(reading is doable by using `raw_exec`).

* fail fast on erroneous SQLite database name

* remove unnecessary wrappers

* open & double close test use the new functions

* only do one copy for BLOB values

(was two before; now we just copy from SQLite to rusqlite and then
reference that and make an Erlang binary pointing to the rusqlite
value).

In the process `XqliteVal` was deemed no longer necessary because we had
to fragment the encoding logic. Thus, having an `impl Encoder` for one
object but still pattern-match on rusqlite Value before invoking the
encoding implementation became a hurdle. This is addressed by making a
simple non-trait-bound function that encodes all possible SQLite values
to Erlang terms.

* BUGFIX: open DB was always overwriting the pool

due to not using the DB path anywhere; it was just making a new pool
every time. This has been fixed to use the DB path as a key in the
global map that keeps connections pools for different databases.

* opening DB now supports options

* rename for clarification

* add Elixir tests exercising open / close / exec

and pragma writing. Mostly check if options are properly parsed and if
errors are what we expect. Exercise the happy paths as well.

* add function to write pragma and return result of the call

* no error if pragma writing returns no value

* support in-memory and temporary databases

which introduced the need for an enum to represent a connection.

* allow parameter list, not just named parameters

* remove all traces of Rust pooling

as we will delegate pooling to the Elixir code (likely via DBConnection)

* remove all traces of Rust pooling

as we will delegate pooling to the Elixir code (likely via DBConnection)

* make clippy happy

* use a better error when parsing (keyword) lists

* maintain dirtyio nif scheduling!

* move function to other helpers

* `:fun` atom -> `:function`

* remove now-unnecessary crates

* upgrade rusqlite

* remote Rust pool options from tests

* 🎉 remove old implementation

* move all code in a single `lib.rs` file

* remove unnecessary atoms

* minimize

* remove unnecessary extra module

* remove references to replaced module

* change Elixir-side pragma implementation

* reintroduce guards but in the central module

* add `.iex.exs` for dev convenience

* remove HTML-parsing code for SQLite PRAGMAs

as this is out of scope for this project.

* juggle comments

* add ad-hoc tests as a fail-safe

* remove unnecessary database name constant

* remove unused function

* remove IO.inspect

* do away with one-character variable names

* remove unused helpers

* async tests

* remove unused helpers

* do away with one-character variable names

* renames

* clarify why we need another module for helpers

* clarify why we need another module for helpers

* do not have two name() types

* remove superfluous separator comments

* too much guards; specs are enough

* remove unneeded types

* make guards clearer

* we do not need supervision for this library

* use rustler atoms, do not redefine them with ours

* replace removed type

* raw_exec -> raw_query as we'll have write version

in a future commit. Also make sure to always return column names and a
number of rows. Rework some Elixir code to accommodate for the Rust
changes.

* add function for INSERT / DELETE / UPDATE

that returns affected rows. Also changed some Elixir code to accommodate
for it.

* add transaction functions + Elixir tests

* add Rust helper to wrap usage of connection

in order to DRY up the code a little, especially to avoid the many
`.map_err` function calls.

* _much_ more detailed errors

* use modern SQLite features

* bump version to 0.2.0

---------

Co-authored-by: Benedikt Reinartz <filmor@gmail.com>